requirements.txt
Keras==2.7.0
numpy==1.21.5
nlpcube==0.3.1.1
gym==0.21.0
gym-anytrading==1.2.0
stable-baselines3==1.3.0
tensorflow==2.7.0
tensorflow-gpu==2.7.0